+++ spacy_conll/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.1.0"
+__version__ = "3.0.0-rc1"
 
 from .formatter import ConllFormatter
 from .utils import init_parser