+++ spacy_conll/utils.py
@@ -5,7 +5,84 @@ from spacy.language import Language
 from spacy.tokens import Doc
 from spacy.vocab import Vocab
 
-from . import ConllFormatter
+
+try:
+    import pandas as pd
+
+    PD_AVAILABLE = True
+except ImportError:
+    PD_AVAILABLE = False
+
+try:
+    import spacy_stanza
+
+    STANZA_AVAILABLE = True
+except ImportError:
+    STANZA_AVAILABLE = False
+
+try:
+    import spacy_udpipe
+
+    UDPIPE_AVAILABLE = True
+except ImportError:
+    UDPIPE_AVAILABLE = False
+
+
+def init_parser(
+    model_or_lang: str,
+    parser: str,
+    *,
+    is_tokenized: bool = False,
+    disable_sbd: bool = False,
+    parser_opts: Optional[Dict] = None,
+    **kwargs,
+) -> Language:
+    """Initialise a spacy-wrapped parser given a language or model and some options.
+    :param model_or_lang: language model to use (must be installed for spaCy but will be automatically downloaded for
+           stanza and UDPipe)
+    :param parser: which parser to use. Parsers other than 'spacy' need to be installed separately. Valid options are
+           'spacy', 'stanza', 'udpipe'. Note that the spacy-* wrappers of those libraries need to be
+           installed, e.g. spacy-stanza
+    :param is_tokenized: indicates whether your text has already been tokenized (space-seperated). When using 'spacy',
+           this option also disabled sentence segmentation completely. For stanza, sentence segmentation will *only*
+           to be done by splitting on new lines. See the documentation:
+           https://stanfordnlp.github.io/stanza/tokenize.html. This optioon does not affect UDPipe.
+    :param disable_sbd: disables spaCy automatic sentence boundary detection (only works for spaCy)
+    :param parser_opts: will be passed to the core pipeline. For spacy, it will be passed to its
+           `.load()` initialisations, for stanza `pipeline_opts` is passed to its `.load_pipeline()`
+           initialisations. UDPipe does not have any keyword arguments
+    :param kwargs: options to be passed to the ConllFormatter initialisation
+    :return: an initialised Language object; the parser
+    """
+    parser_opts = {} if parser_opts is None else parser_opts
+
+    if parser == "spacy":
+        exclude = ["senter", "sentencizer"] if disable_sbd or is_tokenized else []
+        nlp = spacy.load(model_or_lang, exclude=exclude, **parser_opts)
+        if is_tokenized:
+            nlp.tokenizer = SpacyPretokenizedTokenizer(nlp.vocab)
+        if disable_sbd or is_tokenized:
+            nlp.add_pipe("disable_sbd", before="parser")
+    elif parser == "stanza":
+        import spacy_stanza
+        import stanza
+
+        verbose = parser_opts.get("verbose", False)
+        stanza.download(model_or_lang, verbose=verbose)
+        nlp = spacy_stanza.load_pipeline(
+            model_or_lang, verbose=verbose, tokenize_pretokenized=is_tokenized, **parser_opts
+        )
+    elif parser == "udpipe":
+        import spacy_udpipe
+
+        spacy_udpipe.download(model_or_lang)
+        nlp = spacy_udpipe.load(model_or_lang)
+    else:
+        raise ValueError("Unexpected value for 'parser'. Options are: 'spacy', 'stanza', 'udpipe'")
+
+    nlp.add_pipe("conll_formatter", config=kwargs, last=True)
+
+    return nlp
 
 
 class SpacyPretokenizedTokenizer:
@@ -32,66 +109,16 @@ class SpacyPretokenizedTokenizer:
             raise ValueError("Unexpected input format. Expected string to be split on whitespace, or list of tokens.")
 
 
-def init_parser(
-    parser: str = "spacy",
-    model_or_lang: str = "en",
-    *,
-    is_tokenized: bool = False,
-    disable_sbd: bool = False,
-    parser_opts: Optional[Dict] = None,
-    **kwargs,
-) -> Language:
-    """Initialise a spacy-wrapped parser given a language or model and some options.
-    :param parser: which parser to use. Parsers other than 'spacy' need to be installed separately. Valid options are
-           'spacy', 'stanfordnlp', 'stanza', 'udpipe'. Note that the spacy-* wrappers of those libraries need to be
-           installed, e.g. spacy-stanza. Defaults to 'spacy'
-    :param model_or_lang: language model to use (must be installed). Defaults to an English model
-    :param is_tokenized: indicates whether your text has already been tokenized (space-seperated). For stanza and
-           stanfordnlp, this will also cause sentence segmentation *only* to be done by splitting on new lines.
-           See the documentation: https://stanfordnlp.github.io/stanfordnlp/tokenize.html
-           See the documentation: https://stanfordnlp.github.io/stanza/tokenize.html
-    :param disable_sbd: disables spaCy automatic sentence boundary detection (only works for spaCy)
-    :param parser_opts: will be passed to the core pipeline. For spacy and udpipe, it will be passed to their
-           `.load()` initialisations, for stanfordnlp and stanza `pipeline_opts` is passed to to their `.Pipeline()`
-           initialisations
-    :param kwargs: options to be passed to the ConllFormatter initialisation
-    :return: an initialised Language object; the parser
-    """
-    parser_opts = {} if parser_opts is None else parser_opts
+@Language.factory("disable_sbd")
+class SpacyDisableSentenceSegmentation:
+    """Disables spaCy's dependency-based sentence boundary detection. In addition, senter and sentencizer components
+    need to be disabled as well."""
 
-    if parser == "spacy":
-        nlp = spacy.load(model_or_lang, **parser_opts)
-        if is_tokenized:
-            nlp.tokenizer = SpacyPretokenizedTokenizer(nlp.vocab)
-        if disable_sbd:
-            nlp.add_pipe(_prevent_sbd, name="prevent-sbd", before="parser")
-    elif parser == "stanfordnlp":
-        import stanfordnlp
-        from spacy_stanfordnlp import StanfordNLPLanguage
+    def __init__(self, nlp: Language, name: str):
+        self.nlp = nlp
+        self.name = name
 
-        snlp = stanfordnlp.Pipeline(lang=model_or_lang, tokenize_pretokenized=is_tokenized, **parser_opts)
-        nlp = StanfordNLPLanguage(snlp)
-    elif parser == "stanza":
-        import stanza
-        from spacy_stanza import StanzaLanguage
-
-        snlp = stanza.Pipeline(lang=model_or_lang, tokenize_pretokenized=is_tokenized, **parser_opts)
-        nlp = StanzaLanguage(snlp)
-    elif parser == "udpipe":
-        import spacy_udpipe
-
-        nlp = spacy_udpipe.load(model_or_lang, **parser_opts)
-    else:
-        raise ValueError("Unexpected value for 'parser'. Options are: 'spacy', 'stanfordnlp', 'stanza', 'udpipe'")
-
-    conllformatter = ConllFormatter(nlp, **kwargs)
-    nlp.add_pipe(conllformatter, last=True)
-
-    return nlp
-
-
-def _prevent_sbd(doc: Doc):
-    """Disables spaCy's sentence boundary detection."""
-    for token in doc:
-        token.is_sent_start = False
-    return doc
+    def __call__(self, doc: Doc) -> Doc:
+        for token in doc:
+            token.is_sent_start = False
+        return doc